move cache to own file
