+++ src/status.py
@@ -1,27 +1,11 @@
-import json
 from mcstatus import JavaServer
 
-
-class Data:
-
-    max_players: int = 0
-    online_players: int = 0
-    version: str = None
-    active: bool = False
-
-    def __init__(self, max_players, online_players, version, active):
-        self.max_players = max_players
-        self.online_players = online_players
-        self.version = version
-        self.active = active
-
-    def to_json(self):
-        return json.dumps(self, default=lambda o: o.__dict__)
+from src.data_cache import DataCache
 
 
 class Status:
 
-    data: Data = Data(0, 0, None, False)
+    data: DataCache = DataCache(0, 0, None, False)
     cache: str = None
     updated = False
